Template tags library bug fixed
