+++ chunks/templatetags/chunks.py
@@ -1,13 +1,21 @@
 from django import template
-from django.db import models
+from django.conf import settings
 from django.core.cache import cache
-from templatetags import get_cache_suffix
+from django.db import models
+from md5 import md5 
 
 register = template.Library()
 
 Chunk = models.get_model('chunks', 'chunk')
 CACHE_PREFIX = 'chunks_'
 CACHE_SUFFIX = get_cache_suffix()
+
+def get_cache_suffix():
+    try:
+        project_hash = md5(settings.SECRET_KEY).hexdigest()
+        return "_%s" % project_hash
+    except AttributeError:
+        return "_%s" % md5().hexdigest()
 
 def do_get_chunk(parser, token):
     # split_contents() knows not to split quoted strings.
